chore: separate consumer into own file
